Fix handling of duplicate extensions
GitHub includes several languages that have the same extensions.
For example, `.cake` is used for CoffeeScript and for C#.
But which one is used for ` ```cake `?
The answer is complex.
But this commit solves it, it removes duplicate extensions from
languages that will not get highlighted as such if you use them.
Practically, that means `.cake` is no longer present here as an
extension for CoffeeScript.
Previously, which language was chosen was decided based on in which
order languages were loaded: later languages “won”.

There are some cases, where the common extensions, such as `.md`,
is *not* used to highlight as markdown, which one might expect.
It’s instead highlighted as Lisp.
That’s unfortunate, but it’s better to match how GitHub works, and
these cases can be fixed upstream in `github/linguist`.

Finally, an almost never used feature of fenced code blocks on GitHub,
is that you can write explicit extensions (` ```.js `).
Sometimes, when there is a dot, it maps to different languages.
This project now matches GitHub in what values without a dot, and with a
dot, map to.

More info in:
<https://github.com/github/linguist/pull/6338>.