Fix missing `.d.ts` files
Some language files are missing `.d.ts` files.
For performance reasons, the `lang/` folder was ecluded from the
`tsconfig.json` file.
But every file that was *used* in the exports got generated anyway.

Anyway, there are other ways to improve performance, and this does that.

This affects languages that were renamed (such as `source.gfm` which is now
`text.md`) and languages that are optional dependencies (such as
`hidden.manref`).

Closes GH-21.