Replace vscode-oniguruma with oniguruma-to-es
This replaces the dependency on the oniguruma wasm build.
The code is a modified version of
https://github.com/shikijs/shiki/blob/main/packages/engine-javascript.
The upstream has an option `forgiving`, but we don’t. We can add it or
remove it. We may also be able to increase the scope in which the cache
is used. It can be global or per starry-night instance.